+++ tests/Formatting/EmptyTest.php
@@ -11,6 +11,7 @@ use Arokettu\Json5\Values\InlineList;
 use Arokettu\Json5\Values\InlineObject;
 use Arokettu\Json5\Values\ObjectValue;
 use PHPUnit\Framework\TestCase;
+use stdClass;
 
 class EmptyTest extends TestCase
 {
@@ -26,5 +27,19 @@ class EmptyTest extends TestCase
         self::assertEquals("{}\n", Json5Encoder::encode(new ObjectValue([])));
         self::assertEquals("{}\n", Json5Encoder::encode(new InlineObject([])));
         self::assertEquals("{}\n", Json5Encoder::encode(new CompactObject([])));
+    }
+
+    public function testSecondLevelEmptyList(): void
+    {
+        self::assertEquals("[\n    [],\n]\n", Json5Encoder::encode([[]]));
+        self::assertEquals("[\n    [],\n]\n", Json5Encoder::encode([new InlineList([])]));
+        self::assertEquals("[\n    [],\n]\n", Json5Encoder::encode([new CompactList([])]));
+    }
+
+    public function testSecondLevelEmptyObject(): void
+    {
+        self::assertEquals("[\n    {},\n]\n", Json5Encoder::encode([new stdClass()]));
+        self::assertEquals("[\n    {},\n]\n", Json5Encoder::encode([new InlineObject(new stdClass())]));
+        self::assertEquals("[\n    {},\n]\n", Json5Encoder::encode([new CompactObject(new stdClass())]));
     }
 }